+++ src/apps/backend/modules/access-token/rest-api/access-token-router.ts
@@ -8,7 +8,17 @@ export default class AccessTokenRouter {
     const router = Router();
 
     router.post('/', AccessTokenController.createAccessToken);
-    router.post('/phoneLogin', AccessTokenController.createPhoneAccessToken)
+    router.post('/phoneLogin', AccessTokenController.loginWithPhoneNumber);
+
+    router.post(
+      '/phoneLoginVerify',
+      AccessTokenController.verifyWithPhoneNumber,
+    );
+
+    router.post(
+      '/phoneLoginToken',
+      AccessTokenController.createPhoneAccessToken,
+    );
 
     return router;
   }
