+++ src/apps/backend/modules/access-token/types.ts
@@ -18,6 +18,10 @@ export type CreatePhoneAccessTokenParams = {
   phoneNumber: string;
 };
 
+export type PhoneAccountSearchParams = {
+  phoneNumber: string;
+};
+
 export enum AccessTokenErrorCode {
   UNAUTHORIZED_ACCESS = 'ACCESS_TOKEN_ERR_01',
   ACCESS_TOKEN_EXPIRED = 'ACCESS_TOKEN_ERR_02',
