+++ src/apps/backend/modules/account/account-service.ts
@@ -1,7 +1,12 @@
 // import AccountReader from './internal/account-reader';
 import AccountReader from './internal/account-reader';
 import AccountWriter from './internal/account-writer';
-import { Account, AccountSearchParams, CreateAccountParams, phoneAccountSearchParams } from './types';
+import {
+  Account,
+  AccountSearchParams,
+  CreateAccountParams,
+  PhoneAccountSearchParams,
+} from './types';
 
 export default class AccountService {
   public static async createAccount(
@@ -20,7 +25,11 @@ export default class AccountService {
     return AccountReader.getAccountByUsernamePassword(params);
   }
 
-  public static async getAccountByPhone(params:phoneAccountSearchParams) {
+  public static async checkPhoneNumberNotExists(phoneNumber: string) {
+    return AccountReader.checkPhoneNumberNotExists(phoneNumber);
+  }
+
+  public static async getAccountByPhone(params: PhoneAccountSearchParams) {
     return AccountReader.getAccountByPhone(params);
   }
 }
