+++ src/apps/backend/modules/account/internal/account-reader.ts
@@ -6,7 +6,7 @@ import {
   AccountWithUserNameExistsError,
   InvalidCredentialsError,
   PhoneAccountNotFoundError,
-  phoneAccountSearchParams,
+  PhoneAccountSearchParams,
 } from '../types';
 
 import AccountUtil from './account-util';
@@ -25,7 +25,7 @@ export default class AccountReader {
     return AccountUtil.convertAccountDBToAccount(dbAccount);
   }
 
-  public static async getAccountByPhone(params: phoneAccountSearchParams) {
+  public static async getAccountByPhone(params: PhoneAccountSearchParams) {
     const { phoneNumber } = params;
     const dbAccount = await AccountRepository.phoneAccountDB.findOne({
       phoneNumber,
