+++ src/apps/backend/modules/account/rest-api/account-controller.ts
@@ -1,13 +1,8 @@
 import { NextFunction, Request, Response } from 'express';
 
 import AccountService from '../account-service';
-import {
-  Account,
-  CreateAccountParams,
-  phoneAccountSearchParams,
-} from '../types';
+import { Account, CreateAccountParams } from '../types';
 import SMSService from '../../communication/sms-service';
-import AccountReader from '../internal/account-reader';
 
 export default class AccountController {
   public static async createAccount(
@@ -26,22 +21,6 @@ export default class AccountController {
     }
   }
 
-  public static async loginWithPhoneNumber(
-    req: Request,
-    res: Response,
-    next: NextFunction,
-  ): Promise<void> {
-    try {
-      const { phoneNumber } = req.body;
-      const params: phoneAccountSearchParams = { phoneNumber };
-      await AccountReader.getAccountByPhone(params);
-      await SMSService.sendOtp(phoneNumber);
-      res.status(201).send(`otp sent to ${phoneNumber}`);
-    } catch (e) {
-      next(e);
-    }
-  }
-
   public static async createAccountWithPhoneNumber(
     req: Request,
     res: Response,
@@ -49,29 +28,9 @@ export default class AccountController {
   ): Promise<void> {
     try {
       const { phoneNumber } = req.body;
-      await AccountReader.checkPhoneNumberNotExists(phoneNumber);
+      await AccountService.checkPhoneNumberNotExists(phoneNumber);
       await SMSService.sendOtp(phoneNumber);
       res.status(201).send(`otp sent to ${phoneNumber}`);
-    } catch (e) {
-      next(e);
-    }
-  }
-
-  public static async verifyWithPhoneNumber(
-    req: Request,
-    res: Response,
-    next: NextFunction,
-  ): Promise<void> {
-    try {
-      const { phoneNumber, otp } = req.body;
-      const params = { phoneNumber, otp };
-      const response = await SMSService.verifyOtp(params);
-
-      if (response.status === 'approved') {
-        res.status(201).send(`verified successfully ${phoneNumber}`);
-      } else {
-        res.status(422).send(`Incorrect otp try again`);
-      }
     } catch (e) {
       next(e);
     }
